+++ react/components/ProductAvailability.tsx
@@ -1,48 +1,65 @@
-import React, { FunctionComponent, memo } from 'react'
-import classNames from 'classnames'
-
-import styles from '../styles.css'
+import React, { memo } from 'react'
+import type { ProductTypes } from 'vtex.product-context'
 
 import LowStock from './LowStock'
 import HighStock from './HighStock'
+import ShowAvailable from './ShowAvailable'
+import Container from './Container'
 
 interface Props {
   threshold: number
-  lowStockMessage: string
-  highStockMessage: string
-  availableQuantity: number | null | undefined
+  lowStockMessage?: string
+  highStockMessage?: string
+  showAvailability?: 'disabled' | 'stock'
+  showAvailabilityMessage?: string
+  availableQuantity?: ProductTypes.CommercialOffer['AvailableQuantity']
 }
 
-const Container: FunctionComponent = ({ children }) => {
-  return (
-    <div className={classNames(styles.container, 'flex pv2')}>
-      {children}
-    </div>
-  )
-}
-
-const ProductAvailability: FunctionComponent<Props> = ({ threshold, lowStockMessage, highStockMessage, availableQuantity }) => {
-  if (availableQuantity == null || availableQuantity < 1) {
+function ProductAvailability({
+  threshold,
+  lowStockMessage,
+  highStockMessage,
+  showAvailability = 'disabled',
+  showAvailabilityMessage,
+  availableQuantity,
+}: Props) {
+  if (!availableQuantity || availableQuantity < 1) {
     return null
   }
 
   const isLowStock = availableQuantity < threshold
+
+  if (showAvailability === 'stock' && showAvailabilityMessage) {
+    return (
+      <Container>
+        <ShowAvailable
+          text={showAvailabilityMessage}
+          availableQuantity={availableQuantity}
+        />
+      </Container>
+    )
+  }
+
   if (isLowStock && lowStockMessage) {
     return (
       <Container>
-        <LowStock text={lowStockMessage} availableQuantity={availableQuantity} />
+        <LowStock
+          text={lowStockMessage}
+          availableQuantity={availableQuantity}
+        />
       </Container>
     )
   }
-  
-  if (highStockMessage) {
-    return (
-      <Container>
-        <HighStock text={highStockMessage} />
-      </Container>
-    )
+
+  if (!highStockMessage) {
+    return null
   }
-  return null
+
+  return (
+    <Container>
+      <HighStock text={highStockMessage} />
+    </Container>
+  )
 }
 
 export default memo(ProductAvailability)