+++ react/typings/vtex.product-context.d.ts
@@ -1,4 +1,0 @@
-declare module 'vtex.product-context' {
-  import { Context } from 'react'
-  export const ProductContext: Context<any>
-}